Add fleet edition
