+++ libs/ui-components/src/utils/patch.ts
@@ -1,6 +1,7 @@
 import { PatchRequest } from '@flightctl/types';
 import isNil from 'lodash/isNil';
 import uniq from 'lodash/uniq';
+import { FlightCtlLabel } from '../types/extraTypes';
 
 export const appendJSONPatch = <V = unknown>({
   patches,
@@ -65,6 +66,50 @@ export const getStringListPatches = (
       });
     }
   }
+
+  return patches;
+};
+
+export const getLabelPatches = (basePath: string, currentLabels: Record<string, string>, newLabels: FlightCtlLabel[]) => {
+  const patches: PatchRequest = [];
+
+  const allKeys: string[] = Object.entries(currentLabels).map(([key]) => key);
+  newLabels.forEach((newLabel) => {
+    if (!allKeys.includes(newLabel.key)) {
+      allKeys.push(newLabel.key);
+    }
+  });
+
+
+  // TODO do at the higher level object?
+  allKeys.forEach((labelKey) => {
+    const newLabel = newLabels.find((newLabel) => newLabel.key === labelKey);
+
+    // Valueless labels need to be considered separately - "appendJSONPatch" identifies no value as deleted entry
+    const valuelessLabelRemoved = labelKey in currentLabels && !currentLabels[labelKey] && !newLabel;
+    const labelIsNowValueless = labelKey in currentLabels && currentLabels[labelKey] && newLabel && !newLabel.value;
+
+    const path = `${basePath}/${labelKey}`;
+    if (valuelessLabelRemoved) {
+      patches.push({
+        path,
+        op: 'remove',
+      });
+    } else if (labelIsNowValueless) {
+      patches.push({
+        path,
+        op: 'replace',
+        value: '',
+      });
+    } else {
+      appendJSONPatch({
+        path,
+        patches,
+        newValue: newLabel?.value || '',
+        originalValue: currentLabels[labelKey] || '',
+      });
+    }
+  });
 
   return patches;
 };
